requirements.txt
django
djangorestframework
mysqlclient
argon2-cffi
opencv-contrib-python==4.5.3.56
pip install opencv-python==4.6.0.66 --verbose
pillow
numpy
scipy
cython
matplotlib
scikit-image
scikit-learn
tensorflow>=1.3.0
keras>=2.0.8
h5py==2.10.0
imgaug
IPython[all]